Remove common-labels
Fixes #3581

This common-labels functionality made the CLI tool harder to use when
piping the output into other programs. The common labels would be printed
on STDERR, while the rest of the labels and log-lines were emitted on
STDOUT.
I don't think we need this functionality any more, now that we have the
`--exclude-label` option.